improving MAGEMin general reorganisation of the database
