+++ src/MAGEMin.c
@@ -1,37 +1,30 @@
-/**
-Mineral Assemblage Gibbs Energy Minimization		  
---------------------------------------------
+/*@ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+ **
+ **   Project      : MAGEMin
+ **   License      : GNU GENERAL PUBLIC LICENSE Version 3, 29 June 2007
+ **   Developers   : Nicolas Riel, Boris Kaus
+ **   Contributors : Dominguez, H., Green E., Berlie N., and Rummel L.
+ **   Organization : Institute of Geosciences, Johannes-Gutenberg University, Mainz
+ **   Contact      : nriel[at]uni-mainz.de, kaus[at]uni-mainz.de
+ **
+ ** ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ @*/
 
-Contributors: 
-
-- Main developers: Riel N., Kaus. B.
-- Database translation and debugging: Green E., Berlie N., and Rummel L. 
-     
-Contacts: nriel[at]uni-mainz.de, kaus[at]uni-mainz.de 		 
-                                                                                                  
-MAGEMin is written as a parallel C library callable from any petrological/geodynamic tool. For a given set of pressure, temperature and bulk-rock composition MAGEMin uses a combination of linear programming, extended Partitioning Gibbs free Energy and gradient-based local minimization to compute the most stable mineral assemblage     
-      
-Available thermodynamic dataset                       
-================================
+// Available thermodynamic dataset                       
+// ================================
  
-Igneous thermodynamic dataset
-*****************************
-                    
-- Holland et al., 2018 (see http://hpxeosandthermocalc.org)
-- K2O-Na2O-CaO-FeO-MgO-Al2O3-SiO2-H2O-TiO2-O-Cr2O3 chemical system
-- Equations of state for
-	- Pure stoichiometric phases quartz (q), cristobalite (crst), tridymite (trd), coesite (coe), stishovite (stv), kyanite (ky), sillimanite (sill), andalusite (and), rutile (ru) and sphene (sph). 
-	- Solution phases spinel (spn), biotite (bi), cordierite (cd), clinopyroxene (cpx), orthopyroxene (opx), epidote (ep), garnet (g), hornblende (hb), ilmenite (ilm), silicate melt (liq), muscovite (mu), olivine (ol), ternary feldspar (fsp), and aqueous fluid (fl).
-                                                    
-Imported libraries                       
-==================
+// Metapelite (White et al., 2014)
+// Metabasite (Green et al., 2016)
+// Igneous 	  (Holland et al., 2018)
+// Ultramafic (Evans & Frost, 2021)
+                                              
+// Imported libraries                       
+// ==================
 
-- LAPACKE (C version of LAPACK)                         
-- NLopt  (https://nlopt.readthedocs.io/)                
-- uthash (https://troydhanson.github.io/uthash/)        
-- ketopt (https://github.com/attractivechaos/klib/blob/master/ketopt.h) 
-   
- */
+// - LAPACKE (C version of LAPACK)                         
+// - NLopt  (https://nlopt.readthedocs.io/)                
+// - uthash (https://troydhanson.github.io/uthash/)        
+// - ketopt (https://github.com/attractivechaos/klib/blob/master/ketopt.h) 
+
 
 #include <math.h>
 #include <stdio.h>
@@ -40,8 +33,9 @@ Imported libraries
 #include <complex.h> 
 
 #include "uthash.h"
+#include "hash_init.h"
 #include "ketopt.h"
-#include "nlopt.h"                  // requires specifying this in the makefile
+#include "nlopt.h"                  
 #include "mpi.h"
 
 #include "toolkit.h"
@@ -50,7 +44,6 @@ Imported libraries
 
 #include "simplex_levelling.h"
 #include "initialize.h"
-// #include "all_init_database.h"
 #include "ss_min_function.h"
 #include "pp_min_function.h"
 #include "dump_function.h"
@@ -58,6 +51,8 @@ Imported libraries
 #include "phase_update_function.h"
 #include "all_solution_phases.h"
 #include "MAGEMin.h"
+
+
 
 /** 
   Main routine
@@ -847,33 +842,33 @@ Databases InitializeDatabases(	global_variable gv,
 	DB.FS_names  =	get_FS_DB_names(		gv									);
 
 	/* Create endmember Hashtable */
-	struct EM2id *p_s, *tmp_p;
-	struct EM_db EM_return;
+	EM2id *p_s, *tmp_p;
+	EM_db EM_return;
 	int n_em_db = gv.n_em_db;
     for (int i = 0; i < n_em_db; ++i) {
 		char EM_name[20];
-        p_s = (struct EM2id *)malloc(sizeof *p_s);
+        p_s = (EM2id *)malloc(sizeof *p_s);
         strcpy(p_s->EM_tag, DB.EM_names[i]);
         p_s->id = i;
         HASH_ADD_STR( EM, EM_tag, p_s );
     }
 
 	/* Create pure-phase hashtable */
-	struct PP2id *pp_s, *tmp_pp;
+	PP2id *pp_s, *tmp_pp;
     for (int i = 0; i < sizeof(gv.PP_list); ++i) {
-        pp_s = (struct PP2id *)malloc(sizeof *pp_s);
+        pp_s = (PP2id *)malloc(sizeof *pp_s);
         strcpy(pp_s->PP_tag, gv.PP_list[i]);
         pp_s->id = i;
         HASH_ADD_STR( PP, PP_tag, pp_s );
     }
 
 	/* Create fluid species Hashtable */
-	struct FS2id *fs_s, *tmp_fs;
-	struct FS_db FS_return;
+	FS2id *fs_s, *tmp_fs;
+	FS_db FS_return;
 	int n_fs_db = gv.n_fs_db;
     for (int i = 0; i < n_fs_db; ++i) {
 		char EM_name[20];
-        fs_s = (struct FS2id *)malloc(sizeof *fs_s);
+        fs_s = (FS2id *)malloc(sizeof *fs_s);
         strcpy(fs_s->FS_tag, DB.FS_names[i]);
         fs_s->id = i;
         HASH_ADD_STR( FS, FS_tag, fs_s );
