+++ src/gem_function.c
@@ -1,3 +1,13 @@
+/*@ ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+ **
+ **   Project      : MAGEMin
+ **   License      : GNU GENERAL PUBLIC LICENSE Version 3, 29 June 2007
+ **   Developers   : Nicolas Riel, Boris Kaus
+ **   Contributors : Dominguez, H., Green E., Berlie N., and Rummel L.
+ **   Organization : Institute of Geosciences, Johannes-Gutenberg University, Mainz
+ **   Contact      : nriel[at]uni-mainz.de, kaus[at]uni-mainz.de
+ **
+ ** ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ @*/
 /**
   Function to calculate chemical potential of endmembers/pure phases  
 */
@@ -603,7 +613,7 @@ PP_ref G_EM_function(		int 		 EM_database,
 							char		*state			
 ){
 	/* Get thermodynamic data */
-	struct EM_db EM_return;
+	EM_db EM_return;
 	int i, p_id = find_EM_id(name);
 	EM_return   = Access_EM_DB(p_id, EM_database);
 	
@@ -932,7 +942,7 @@ PP_ref G_FS_function(		int 		 len_ox,
 
 
 	/* Get thermodynamic data */
-	struct FS_db FS_return;
+	FS_db FS_return;
 	int i, p_id = find_FS_id(name);
 	FS_return   = Access_FS_DB(p_id);
